testXray/requirements.txt
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
pytest-base-url
allure-pytest
pyyaml
allure-pytest